requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0
psycopg2>=2.8
gunicorn>=20.0  # WSGI HTTP Server to run Django in production
Pillow